+++ src/query/mod.rs
@@ -29,7 +29,7 @@ fn parse_args(abbr: &OsStr) -> Result<(PathBuf, Vec<Abbr>)> {
         return Err(Error::EmptyAbbr);
     }
 
-    let (start_path, suffix) = extract_start_path(as_path(abbr));
+    let (start_path, suffix) = decompose_arg(as_path(abbr))?;
 
     let start_path = start_path
         .map(|path| Ok(path))
@@ -41,7 +41,7 @@ fn parse_args(abbr: &OsStr) -> Result<(PathBuf, Vec<Abbr>)> {
             component
                 .as_os_str()
                 .to_str()
-                .ok_or(Error::ArgInvalidUnicode)
+                .ok_or(Error::InvalidUnicode)
                 .and_then(|s| Abbr::from_string(s.to_string()))
         })
         .collect::<Result<Vec<Abbr>>>()?;
@@ -53,30 +53,55 @@ fn parse_args(abbr: &OsStr) -> Result<(PathBuf, Vec<Abbr>)> {
     Ok((start_path, abbr))
 }
 
-fn extract_start_path<'p>(
-    arg: &'p Path,
-) -> (Option<PathBuf>, Vec<Component<'p>>) {
-    let mut suffix = arg.components().peekable();
+fn maybe_parse_dots(component: &str) -> Option<u32> {
+    component.chars()
+    	.try_fold(0, |occurences, c| if c == '.' { Some(occurences + 1) } else { None })
+        .and_then(|occurences| if occurences >= 1 {
+            Some(occurences - 1)
+        } else {
+            None
+        })
+}
+
+fn decompose_arg<'a, P>(arg: &'a P) -> Result<(Option<PathBuf>, Vec<Component<'a>>)>
+where P: AsRef<Path> + ?Sized + 'a {
+    use std::path::Component::*;
+
+    let arg = arg.as_ref();
+    let mut arg = arg.components().peekable();
     let mut prefix: Option<PathBuf> = None;
 
-    while let Some(component) = suffix.next_if(|component| {
-        use std::path::Component::*;
-        match component {
-            Prefix(_) | RootDir | CurDir | ParentDir => true,
-            Normal(_) => false,
-        }
-    }) {
+    let mut push_to_prefix = |component| {
         match prefix {
             Some(ref mut prefix) => prefix.push(component),
             None => {
                 prefix = Some(PathBuf::from(as_path(&component)));
             }
         }
+    };
+
+    while let Some(component) = arg.peek() {
+        match component {
+            Prefix(_) | RootDir | CurDir | ParentDir => push_to_prefix(component.clone()),
+            Normal(component_os) => {
+                let component = component_os.to_str().ok_or(Error::InvalidUnicode)?;
+
+                if let Some(n_dots) = maybe_parse_dots(component) {
+                    for _ in 0..n_dots {
+                    	push_to_prefix(ParentDir);
+                    }
+                } else {
+                    break;
+                }
+            },
+        }
+
+        arg.next();
     }
 
-    let suffix = suffix.collect();
+    let arg = arg.collect();
 
-    (prefix, suffix)
+    Ok((prefix, arg))
 }
 
 #[cfg(test)]
@@ -89,26 +114,38 @@ mod test {
     use pretty_assertions::assert_eq;
 
     #[test]
-    fn test_extract_start_path() {
+    fn test_decompose_arg() {
         // No start path.
-        let (start_path, suffix) = extract_start_path(as_path("a/b/c"));
+        let (start_path, suffix) = decompose_arg(as_path("a/b/c")).unwrap();
         let first_abbr = suffix[0].as_os_str();
 
         assert!(start_path.is_none());
         assert_eq!(first_abbr, "a");
 
         // Root dir.
-        let (start_path, suffix) = extract_start_path(as_path("/gn"));
+        let (start_path, suffix) = decompose_arg(as_path("/gn")).unwrap();
         let first_abbr = suffix[0].as_os_str();
 
         assert_eq!(start_path.unwrap(), as_path("/"));
         assert_eq!(first_abbr, "gn");
 
         // Multiple `..` and `.`.
-        let (start_path, suffix) = extract_start_path(as_path(".././../do"));
+        let (start_path, suffix) = decompose_arg(as_path(".././../do")).unwrap();
         let first_abbr = suffix[0].as_os_str();
 
         assert_eq!(start_path.unwrap(), as_path(".././.."));
         assert_eq!(first_abbr, "do");
+
+        // Three or more dots.
+        let (start_path, suffix) = decompose_arg(as_path("./../.../..../oops")).unwrap();
+        let first_abbr = suffix[0].as_os_str();
+
+        // . = 0
+        // .. = 1
+        // ... = 2
+        // .... = 3
+        // total of 6
+        assert_eq!(start_path.unwrap(), as_path("./../../../../../../"));
+        assert_eq!(first_abbr, "oops");
     }
 }